pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "agentune_analyze"
version = "0.1.0"
description = "A library for discovering insights in data"
authors = ["SparkBeyond Team"]
readme = "README.md"
packages = [{include = "agentune.analyze"}]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
duckdb = ">=1.2.1,<2.0.0"
pyarrow = ">=19.0.1"
numpy = ">=2.2.4,<3.0.0"
janus = ">=2.0.0,<3.0.0"
polars = ">=1.27.1,<2.0.0"
attrs = "^25.3.0"
scikit-learn = ">=1.6.1,<2.0"
scipy = ">=1.15.3,<2.0"
scipy-stubs = ">=1.15.3.0"
cattrs = "^25.1.1"
httpx = "^0.28.1"
frozendict = "^2.4.6"
llama-index-core = "^0.12.48"

[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"
ruff = "^0.11.6"
pytest-mypy-plugins = "^3.2.0"
pyarrow-stubs = ">=19.1"
pytest = "^8.3.5"
matplotlib = "^3.10.3"
pytest-asyncio = "^1.0.0"
llama-index-llms-openai = "^0.4.7"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
testpaths = [ "tests" ]
pythonpath = ["agentune", "tests/agentune"]
tmp_path_retention_policy = "none"
